p-multigrid: remove custom interpolation code
