+++ tests/firedrake/multigrid/test_p_multigrid.py
@@ -1,5 +1,4 @@
 import pytest
-import numpy as np
 from firedrake import *
 
 
@@ -72,80 +71,6 @@ def test_reconstruct_degree(tp_mesh, mixed_family):
 
         elist.append(e)
         assert e == PMGPC.reconstruct_degree(elist[0], degree)
-
-
-@pytest.mark.parametrize("family", ["Q", "NCE", "NCF", "DQ"])
-def test_prolong_basic(tp_mesh, family):
-    """ Interpolate a constant function between low-order and high-order spaces
-    """
-    from firedrake.preconditioners.pmg import prolongation_matrix_matfree
-    if tp_mesh.topological_dimension == 2:
-        family = family.replace("N", "RT")
-
-    fs = [FunctionSpace(tp_mesh, family, degree) for degree in (1, 2)]
-    u, v = [Function(V) for V in fs]
-
-    u.assign(1)
-    P = prolongation_matrix_matfree(u, v).getPythonContext()
-    P._prolong()
-    assert np.allclose(v.dat.data, 1)
-
-
-def test_prolong_de_rham(tp_mesh):
-    """ Interpolate a linear vector function between [H1]^d, HCurl and HDiv spaces
-        where it can be exactly represented
-    """
-    from firedrake.preconditioners.pmg import prolongation_matrix_matfree
-
-    tdim = tp_mesh.topological_dimension
-    b = Constant(list(range(tdim)))
-    if tp_mesh.extruded_periodic:
-        expr = b
-    else:
-        mat = diag(Constant([tdim+1]*tdim)) + Constant([[-1]*tdim]*tdim)
-        expr = b + dot(mat, SpatialCoordinate(tp_mesh))
-
-    cell = tp_mesh.ufl_cell()
-    elems = [VectorElement(FiniteElement("Q", cell=cell, degree=2)),
-             FiniteElement("NCE" if tdim == 3 else "RTCE", cell=cell, degree=2),
-             FiniteElement("NCF" if tdim == 3 else "RTCF", cell=cell, degree=2)]
-
-    fs = [FunctionSpace(tp_mesh, e) for e in elems]
-    us = [Function(V) for V in fs]
-    us[0].interpolate(expr)
-    for u in us:
-        for v in us:
-            if u != v:
-                P = prolongation_matrix_matfree(u, v).getPythonContext()
-                P._prolong()
-                assert errornorm(expr, v) < 1E-14
-
-
-def test_prolong_low_order_to_restricted(tp_mesh, tp_family, variant):
-    """ Interpolate a low-order function to interior and facet high-order spaces
-        and ensure that the sum of the two high-order functions is equal to the
-        low-order function
-    """
-    from firedrake.preconditioners.pmg import prolongation_matrix_matfree
-
-    degree = 5
-    cell = tp_mesh.ufl_cell()
-    element = FiniteElement(tp_family, cell=cell, degree=degree, variant=variant)
-    Vi = FunctionSpace(tp_mesh, RestrictedElement(element, restriction_domain="interior"))
-    Vf = FunctionSpace(tp_mesh, RestrictedElement(element, restriction_domain="facet"))
-    Vc = FunctionSpace(tp_mesh, tp_family, degree=1)
-
-    ui = Function(Vi)
-    uf = Function(Vf)
-    uc = Function(Vc)
-    uc.dat.data[0::2] = 0.0
-    uc.dat.data[1::2] = 1.0
-
-    for v in [ui, uf]:
-        P = prolongation_matrix_matfree(uc, v).getPythonContext()
-        P._prolong()
-
-    assert norm(ui + uf - uc, "L2") < 1E-13
 
 
 @pytest.fixture(params=["triangles", "quadrilaterals"], scope="module")
@@ -557,50 +482,3 @@ def test_p_fas_nonlinear_scalar():
     check_coarsen_quadrature(solver_npmg)
     iter_npmg = solver_npmg.snes.getLinearSolveIterations()
     assert 2*iter_pfas <= iter_npmg
-
-
-@pytest.fixture
-def piola_mesh():
-    return UnitDiskMesh(3)
-
-
-@pytest.mark.parametrize("mat_type", ("matfree", "aij"))
-@pytest.mark.parametrize("mixed", (False, True), ids=("standalone", "mixed"))
-@pytest.mark.parametrize("family, degree", (("CG", 4), ("N2curl", 2), ("N1div", 3)))
-def test_pmg_transfer_piola(piola_mesh, family, degree, mixed, mat_type):
-    """Test prolongation and restriction kernels for piola-mapped elements.
-    """
-    from firedrake.preconditioners.pmg import prolongation_matrix_matfree, prolongation_matrix_aij
-    Vf = FunctionSpace(piola_mesh, family, degree)
-    if mixed:
-        DG = FunctionSpace(Vf.mesh(), "DG", 2)
-        Vf = Vf * Vf * DG
-    Vc = Vf.reconstruct(degree=1)
-
-    Vf_bcs = [DirichletBC(Vf.sub(0), 0, "on_boundary")]
-    Vc_bcs = [DirichletBC(Vc.sub(0), 0, "on_boundary")]
-    if mat_type == "matfree":
-        P = prolongation_matrix_matfree(Vc, Vf, Vc_bcs, Vf_bcs)
-    else:
-        P = prolongation_matrix_aij(Vc, Vf, Vc_bcs, Vf_bcs)
-
-    uc = Function(Vc)
-    uf = Function(Vf)
-    with uc.dat.vec_wo as xc:
-        xc.setRandom()
-    for bc in Vc_bcs:
-        bc.zero(uc)
-    with uc.dat.vec_ro as xc, uf.dat.vec as xf:
-        P.mult(xc, xf)
-    assert norm(uf - uc) < 1E-12
-
-    rc = Cofunction(Vc.dual())
-    rf = Cofunction(Vf.dual())
-    with rf.dat.vec_wo as xf:
-        xf.setRandom()
-    for bc in Vf_bcs:
-        bc.zero(rf)
-    with rf.dat.vec_ro as xf, rc.dat.vec as xc:
-        P.multTranspose(xf, xc)
-
-    assert abs(assemble(action(rf, uf)) - assemble(action(rc, uc))) < 1E-11
